swap M2 and MS6.
