+++ manuscripts/method/03.results.md
@@ -13,17 +13,10 @@ The resulting clustering provides coordinated output that can be used in a few d
 
 In parallel or independently, one can interrogate the resulting PSSMs to describe the overall sequence features of that cluster. These outputs can be readily compared to other information such as experimentally generated specificity profiles of putative upstream kinases via Position Specific Scanning Libraries (PSPL), to infer upstream kinases [@PMID:10945990; @PMID:20159853; @PMID:31124786; @PMID:15782149; @PMID:26551075]. We extracted a collection of 62 kinase specificity profiles to find what cluster motifs resemble most accurately the optimal motif of putative upstream kinases (Fig. {@fig:method}C) [@PMID:18765831; @PMID:15782149; @PMID:26551075]. However, kinase-substrate specificity is also dictated by features outside of the immediate substrate region, and we note that our approach is more general than strictly assembling kinase-substrate predictions. Overall, this overview demonstrates how DDMC can take complex, coordinated signaling measurements and find patterns in the phosphorylation signals to reconstruct signaling networks and associate particular clusters and phenotypes.
 
-### Dual data-motif clustering strategy robustly imputes missing values
+### A dual data-motif strategy improves prediction of different phenotypes
+Text
 
-![**Benchmarking the robustness of motif clustering to missing measurements.** A) A schematic of the process for quantifying robustness to missing values. Any peptides containing less than 7 TMT experiments were discarded. For the remaining 15904 peptides, an entire random TMT experiment was removed per peptide and these values were stored for later comparison. Next, these artifical missing values were imputed using either a baseline strategy (peptide mean/minimum signal) or the corresponding cluster center. Once a mean squared error was computed for each peptide, the second iteration repeats this process by removing a second TMT experiment. A total of 5 random TMT experiments per peptide were imputed by clustering using either the peptides' phosphorylation signal only (B), the peptides' sequence only (C) or a mix of both information sources (D). The imputation errors were calculated across different number of clusters (E) and different data-vs-sequence weights (F).](figureM2.svg "Figure 2"){#fig:missingness width="100%"}
-
-To evaluate the robustness of our combinad dual data-motif clustering (DDMC) method in analyzing incomplete data sets, we designed a computational experiment wherein we removed specific observations and attempted to predict them using the cluster centers corresponding to the peptides those missing values belonged to (Figure 2A). The resulting mean squared errors between the actual and predicted values were compared to compared to commonly used imputation strategies such as the peptides' mean or minimum signal. Furthermore, we evaluated the imputation performance of our method when clustering the data by phosphorylation signal only, by sequence similarity only, or by combining both information sources (2B-D). These results indicate that MDCC clearly imputes missing values more accurately than the peptides' mean signal in all three cases. No noticeable differences were observed in the imputations performance using data or sequence only, or a mix of both. Next, we re-implemented this strategy to find out what number of clusters (2E) and weight (2F) yielded more accurate predictions. We found that a larger number of clusters and smaller weights (prioritizing phosphorylation signal over peptide sequence) minimized the imputation error. These data suggests that DDMC outperforms standard imputation strategies such as peptide mean signal (and minimum signal).   
-
-#TODO: 
-- Compare to minimum signal
-- Filling the missingness with 0's
-- Matrix completion by PCA
-- Add legend for each regression line in plots B-F and mix labels/titles
+![**Sequence information improves model prediction.** A) .](figureM2.svg "Figure 2"){#fig:preds width="100%"}
 
 
 ### Shared differences between tumor and normal adjascent tissue
