+++ manuscripts/method/20.supp-figures.md
@@ -33,3 +33,17 @@ DDMC simultaneously models phosphorylation signal and peptide sequence to cluste
 ### Comparing the predictive of performance of using DDMC to different regression strategies
 
 ![**Comparing the predictive performance of a lasso model using different modeling strategies.** (A-D) Lasso's predictive performance of STK11 mutational status using DDMC clusters (A), unclustered signaling data (B), k-means clusters (C) and GMM clusters (D). Predictive performance of a lasso model regressed against EGFRm/ALKf mutational status. Note that a complete portion of the entire data set containing 1311 peptides was used for every modeling strategy other than DDMC. DDMC was fit to 30561 peptides that include at least a minimum of 2 10-plex TMT experiments.](figureMS5.svg){#fig:RegressionsS tag="S5"}
+
+
+<!-- (Supp. Figure 6) -->
+### Dual data-motif clustering strategy robustly imputes missing values
+
+![**Benchmarking the robustness of motif clustering to missing measurements.** A) A schematic of the process for quantifying robustness to missing values. Any peptides containing less than 7 TMT experiments were discarded. For the remaining 15904 peptides, an entire random TMT experiment was removed per peptide and these values were stored for later comparison. Next, these artifical missing values were imputed using either a baseline strategy (peptide mean/minimum signal) or the corresponding cluster center. Once a mean squared error was computed for each peptide, the second iteration repeats this process by removing a second TMT experiment. A total of 5 random TMT experiments per peptide were imputed by clustering using either the peptides' phosphorylation signal only (B), the peptides' sequence only (C) or a mix of both information sources (D). The imputation errors were calculated across different number of clusters (E) and different data-vs-sequence weights (F).](figureMS6.svg){#fig:missingness width="100%". tag="S6"}
+
+To evaluate the robustness of our combinad dual data-motif clustering (DDMC) method in analyzing incomplete data sets, we designed a computational experiment wherein we removed specific observations and attempted to predict them using the cluster centers corresponding to the peptides those missing values belonged to (Figure 2A). The resulting mean squared errors between the actual and predicted values were compared to compared to commonly used imputation strategies such as the peptides' mean or minimum signal. Furthermore, we evaluated the imputation performance of our method when clustering the data by phosphorylation signal only, by sequence similarity only, or by combining both information sources (2B-D). These results indicate that MDCC clearly imputes missing values more accurately than the peptides' mean signal in all three cases. No noticeable differences were observed in the imputations performance using data or sequence only, or a mix of both. Next, we re-implemented this strategy to find out what number of clusters (2E) and weight (2F) yielded more accurate predictions. We found that a larger number of clusters and smaller weights (prioritizing phosphorylation signal over peptide sequence) minimized the imputation error. These data suggests that DDMC outperforms standard imputation strategies such as peptide mean signal (and minimum signal).
+
+#TODO: 
+- Compare to minimum signal
+- Filling the missingness with 0's
+- Matrix completion by PCA
+- Add legend for each regression line in plots B-F and mix labels/titles
